Add code that moves msgs in run to que on quemgr restart

+++ code/test/setup_test_queues.sh
@@ -91,10 +91,24 @@ if [ "$?" -ne "0" ]; then
   exit 1
 fi
 
+echo "Creating the test run queue..."
+curl -0 http://127.0.0.1:${rq_port}/new_queue -sL -F queue[name]=test_run -F queue[script]=./code/test/test_script.sh -F queue[num_workers]=3 -o _install_test_run.txt
+if [ "$?" -ne "0" ]; then
+  echo "Sorry, web server for RQ failed to respond correctly"
+  exit 1
+fi
+
+egrep "successqueue created" _install_test_run.txt > /dev/null
+if [ "$?" -ne "0" ]; then
+  echo "Sorry, system didn't create test_run queue"
+  exit 1
+fi
+
 rm _home.txt
 rm _install.txt
 rm _install_test.txt
 rm _install_test_coalesce.txt
+rm _install_test_run.txt
 
 echo "ALL DONE SUCCESSFULLY"
